перемещает расположение интерфейсов поближе к моделям
