+++ src/app/models/file/index.ts
@@ -1,1 +1,2 @@
 export * from "./file";
+export * from "./types";
