+++ src/app/models/file/types.ts
@@ -1,0 +1,8 @@
+export interface IFile {
+    getName(): string;
+}
+
+export interface IMediaFile extends IFile {
+    getPath(): string;
+    getPoster(): string;
+}
