+++ src/app/models/linked-list/types.ts
@@ -5,3 +5,12 @@ export interface ILinkedItem<T> {
     linkPrev(link: ILinkedItem<T> | null): void;
     getValue(): T;
 }
+
+export interface ILinkedList<T> {
+    add(item: T): T;
+    current(): T;
+    next(): T | null;
+    prev(): T | null;
+    hasNext(): boolean;
+    hasPrev(): boolean;
+}
