+++ src/app/models/player/index.ts
@@ -1,1 +1,2 @@
 export * from "./video-player";
+export * from "./types";
