--- src/interfaces/linked-list/index.ts
@@ -1,2 +1,0 @@
-export * from "./linked-list";
-export * from "./linked-item";
