--- src/interfaces/linked-list/linked-list.ts
@@ -1,10 +1,0 @@
-
-
-export interface ILinkedList<T> {
-    add(item: T): T;
-    current(): T;
-    next(): T | null;
-    prev(): T | null;
-    hasNext(): boolean;
-    hasPrev(): boolean;
-}
